Remove usePopulateProduct hook
