--- src/hooks/usePopulateProduct.js
@@ -1,156 +1,0 @@
-import { gql, useMutation } from "@apollo/client";
-
-const PRODUCTS_QUERY = gql`
-  mutation populateProduct($input: ProductInput!) {
-    productCreate(input: $input) {
-      product {
-        title
-      }
-    }
-  }
-`;
-
-export function usePopulateProduct() {
-  return useMutation(PRODUCTS_QUERY);
-}
-
-export function randomTitle() {
-  const adjective = ADJECTIVES[Math.floor(Math.random() * ADJECTIVES.length)];
-  const noun = NOUNS[Math.floor(Math.random() * NOUNS.length)];
-
-  return `${adjective} ${noun}`;
-}
-
-const ADJECTIVES = [
-  "autumn",
-  "hidden",
-  "bitter",
-  "misty",
-  "silent",
-  "empty",
-  "dry",
-  "dark",
-  "summer",
-  "icy",
-  "delicate",
-  "quiet",
-  "white",
-  "cool",
-  "spring",
-  "winter",
-  "patient",
-  "twilight",
-  "dawn",
-  "crimson",
-  "wispy",
-  "weathered",
-  "blue",
-  "billowing",
-  "broken",
-  "cold",
-  "damp",
-  "falling",
-  "frosty",
-  "green",
-  "long",
-  "late",
-  "lingering",
-  "bold",
-  "little",
-  "morning",
-  "muddy",
-  "old",
-  "red",
-  "rough",
-  "still",
-  "small",
-  "sparkling",
-  "throbbing",
-  "shy",
-  "wandering",
-  "withered",
-  "wild",
-  "black",
-  "young",
-  "holy",
-  "solitary",
-  "fragrant",
-  "aged",
-  "snowy",
-  "proud",
-  "floral",
-  "restless",
-  "divine",
-  "polished",
-  "ancient",
-  "purple",
-  "lively",
-  "nameless",
-];
-
-const NOUNS = [
-  "waterfall",
-  "river",
-  "breeze",
-  "moon",
-  "rain",
-  "wind",
-  "sea",
-  "morning",
-  "snow",
-  "lake",
-  "sunset",
-  "pine",
-  "shadow",
-  "leaf",
-  "dawn",
-  "glitter",
-  "forest",
-  "hill",
-  "cloud",
-  "meadow",
-  "sun",
-  "glade",
-  "bird",
-  "brook",
-  "butterfly",
-  "bush",
-  "dew",
-  "dust",
-  "field",
-  "fire",
-  "flower",
-  "firefly",
-  "feather",
-  "grass",
-  "haze",
-  "mountain",
-  "night",
-  "pond",
-  "darkness",
-  "snowflake",
-  "silence",
-  "sound",
-  "sky",
-  "shape",
-  "surf",
-  "thunder",
-  "violet",
-  "water",
-  "wildflower",
-  "wave",
-  "water",
-  "resonance",
-  "sun",
-  "wood",
-  "dream",
-  "cherry",
-  "tree",
-  "fog",
-  "frost",
-  "voice",
-  "paper",
-  "frog",
-  "smoke",
-  "star",
-];
